Split up testfiles for pethelper too
