+++ src/tests/hoplalib/zoo/test_petmodels.py
@@ -5,7 +5,6 @@ import random
 from hopla.hoplalib.common import GlobalConstants
 from hopla.hoplalib.zoo.petmodels import FeedingStatus, InvalidFeedingStatus, Pet, InvalidPet
 from hopla.hoplalib.zoo.petdata import PetData
-from hopla.hoplalib.zoo.petcontroller import FeedPostRequester
 
 _SAMPLE_SIZE = 50
 """
@@ -13,27 +12,6 @@ Sample size to use for pytest.mark.parameterize when unittests become slow
 """
 
 
-class TestFeedPostRequester:
-    def test_path_is_ok(self):
-        pet = "Wolf-Golden"
-        food = "Honey"
-
-        feed_requester = FeedPostRequester(pet_name=pet,
-                                           food_name=food)
-
-        assert feed_requester.path == f"/user/feed/{pet}/{food}"
-
-    def test_url_is_okay(self):
-        pet = "Wolf-Shimmer"
-        food = "Meat"
-
-        feed_requester = FeedPostRequester(pet_name=pet,
-                                           food_name=food)
-
-        result_url = feed_requester.feed_pet_food_url
-
-        expected_url = f"https://habitica.com/api/v3/user/feed/{pet}/{food}"
-        assert result_url == expected_url
 
 
 class TestPetInit:
